chore: Update App component to include ProductProvider and CartProvider
